+++ src/main/java/org/betonquest/betonquest/compatibility/mythicmobs/MythicMobsIntegrator.java
@@ -38,7 +38,7 @@ public class MythicMobsIntegrator implements Integrator {
         final QuestTypeRegistries questRegistries = plugin.getQuestRegistries();
         questRegistries.getConditionTypes().register("mythicmobdistance", new MythicMobDistanceConditionFactory(apiHelper, data));
         plugin.registerObjectives("mmobkill", MythicMobKillObjective.class);
-        questRegistries.getEventTypes().register("mspawnmob", new MythicSpawnMobEventFactory(apiHelper, data, plugin.getVariableProcessor()));
+        questRegistries.getEventTypes().registerCombined("mspawnmob", new MythicSpawnMobEventFactory(apiHelper, data, plugin.getVariableProcessor()));
         if (Compatibility.getHooked().contains("ProtocolLib")) {
             MythicHider.start();
         }