+++ src/main/java/org/betonquest/betonquest/compatibility/mythicmobs/events/MythicSpawnMobEventFactory.java
@@ -4,6 +4,8 @@ import io.lumine.mythic.bukkit.BukkitAPIHelper;
 import org.betonquest.betonquest.Instruction;
 import org.betonquest.betonquest.api.quest.event.Event;
 import org.betonquest.betonquest.api.quest.event.EventFactory;
+import org.betonquest.betonquest.api.quest.event.StaticEvent;
+import org.betonquest.betonquest.api.quest.event.StaticEventFactory;
 import org.betonquest.betonquest.compatibility.Compatibility;
 import org.betonquest.betonquest.exceptions.InstructionParseException;
 import org.betonquest.betonquest.instruction.variable.VariableNumber;
@@ -11,13 +13,14 @@ import org.betonquest.betonquest.instruction.variable.VariableString;
 import org.betonquest.betonquest.instruction.variable.location.VariableLocation;
 import org.betonquest.betonquest.quest.PrimaryServerThreadData;
 import org.betonquest.betonquest.quest.event.PrimaryServerThreadEvent;
+import org.betonquest.betonquest.quest.event.PrimaryServerThreadStaticEvent;
 import org.betonquest.betonquest.quest.registry.processor.VariableProcessor;
 import org.betonquest.betonquest.utils.Utils;
 
 /**
  * Factory to create {@link MythicSpawnMobEvent}s from {@link Instruction}s.
  */
-public class MythicSpawnMobEventFactory implements EventFactory {
+public class MythicSpawnMobEventFactory implements EventFactory, StaticEventFactory {
     /**
      * Expected format: identifier:amount
      */
@@ -75,5 +78,24 @@ public class MythicSpawnMobEventFactory implements EventFactory {
                 Utils.addPackage(instruction.getPackage(), markedString)
         );
         return new PrimaryServerThreadEvent(new MythicSpawnMobEvent(apiHelper, loc, mob, level, amount, privateMob, targetPlayer, marked), data);
+    }
+
+    @Override
+    public StaticEvent parseStaticEvent(final Instruction instruction) throws InstructionParseException {
+        final VariableLocation loc = instruction.getLocation();
+        final String[] mobParts = instruction.next().split(":");
+        if (mobParts.length != MOB_FORMAT_LENGTH) {
+            throw new InstructionParseException("Wrong mob format");
+        }
+        final String mob = mobParts[0];
+        final VariableNumber level = instruction.getVarNum(mobParts[1]);
+        final VariableNumber amount = instruction.getVarNum();
+        final String markedString = instruction.getOptional("marked");
+        final VariableString marked = markedString == null ? null : new VariableString(
+                variableProcessor,
+                instruction.getPackage(),
+                Utils.addPackage(instruction.getPackage(), markedString)
+        );
+        return new PrimaryServerThreadStaticEvent(new MythicSpawnMobEvent(apiHelper, loc, mob, level, amount, false, false, marked), data);
     }
 }